requirements.txt
dash==2.17.1
dash-bootstrap-components==1.6.0
pandas>=2.0
rsconnect-python
databricks-sql-connector[pyarrow]
databricks-sdk
scikit-learn
dash[diskcache]
dotenv
